backend/requirements.txt
Flask==2.1.0
flask-cors==3.0.10
google-generativeai==0.3.1
python-dotenv==0.19.2
Flask-SQLAlchemy==2.5.1
youtube-transcript-api==0.6.0
